chore:  types 파일 제거
