--- frontend/src/components/ArrowLeftButton/ArrowLeft.tsx
@@ -1,30 +1,0 @@
-import { ArrowLeftProps } from './ArrowLeft.type';
-
-export default function ArrowLeft({ onClick }: ArrowLeftProps) {
-  return (
-    <button onClick={onClick}>
-      <svg
-        width='14'
-        height='15'
-        viewBox='0 0 14 15'
-        fill='none'
-        xmlns='http://www.w3.org/2000/svg'
-      >
-        <path
-          d='M6.99984 11.5832L2.9165 7.49984L6.99984 3.4165'
-          stroke='#4A5565'
-          stroke-width='1.16667'
-          stroke-linecap='round'
-          stroke-linejoin='round'
-        />
-        <path
-          d='M11.0832 7.5H2.9165'
-          stroke='#4A5565'
-          stroke-width='1.16667'
-          stroke-linecap='round'
-          stroke-linejoin='round'
-        />
-      </svg>
-    </button>
-  );
-}
