--- frontend/src/components/ArrowLeftButton/ArrowLeft.type.ts
@@ -1,3 +1,0 @@
-export interface ArrowLeftProps extends React.ComponentProps<'button'> {
-  onClick: () => void;
-}
