--- frontend/src/components/ArrowLeftButton/ArrowLeftButton.type.ts
@@ -1,3 +1,0 @@
-export interface ArrowLeftButtonProps extends React.ComponentProps<'button'> {
-  onClick: () => void;
-}
