--- frontend/src/components/BasicTextArea/BasicTextArea.type.ts
@@ -1,7 +1,0 @@
-export interface BasicTextAreaProps {
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
-  placeholder: string;
-  maxLength?: number;
-  minLength?: number;
-}
