--- frontend/src/components/Header/Header.type.ts
@@ -1,4 +1,0 @@
-export interface HeaderProps {
-  title: string;
-  subtitle: string;
-}
