--- frontend/src/domains/user/Suggestions/components/Banner/Banner.type.ts
@@ -1,4 +1,0 @@
-export interface BannerProps {
-  title: string;
-  description: string;
-}
