--- frontend/src/domains/user/Suggestions/components/SuggestionsFormField/SuggestionsFormField.type.ts
@@ -1,4 +1,0 @@
-export interface SuggestionsFormFieldProps {
-  label: string;
-  children: React.ReactNode;
-}
